restructure AccountPrompt
